lenlen: refactor code to be on chain
