+++ projects/helper/coreAssets.json
@@ -1086,6 +1086,9 @@
   ],
   "vision": [
     "0x1db6cdc620388a0b6046b20cd59503a0839adcff",
+    "0x4dE4B92C12dD4584873e72447573382C41da5597",
+    "0x9261Bb697FE632dd4746c78406E955123085Bf4C",
+    "0xf6515Ce2ab3482C84Daaef961793bdFa3266b8ad",
     "0x79ffbc4fff98b821d59dbd7b33f91a2783006b6f"
   ],
   "kava": [
