+++ projects/lenlen/index.js
@@ -1,15 +1,16 @@
-const axios = require("axios");
 
-async function fetch() {
-  let {data: response} = await axios.get('https://api.v-token.io/api/lenen/defiLlama')
-  let tvl = 0
-  if (response.code === 200 && response.data) {
-    tvl = parseFloat(response.data.total_value_locked ?? '0') + parseFloat(response.data.total_borrow ?? '0')
+const { compoundV3Exports } = require('../helper/compoundV3')
+
+const markets = [
+  "0xc9C1B486bB027cD7023019ADA41F17109eE6c722", // USDT Market
+]
+
+const collaterals = [
+  '0x1db6cdc620388a0b6046b20cd59503a0839adcff',
+]
+
+module.exports = compoundV3Exports({
+  vision: {
+    markets, collaterals,
   }
-  
-  return tvl;
-}
-
-module.exports = {
-  fetch
-}
+})
